Major rework of implements functionality. Some tests not working, don't merge into upstream yet.
The main issue with the previous implementation was that it fought with Delphi's Generics implementation too much. There is a major need for interface generic methods in the way I previously worked this.
Now this has been simplified so that the proxy does not know how to create its child proxies, its just given them. Its the TMock<T> record which knows how to achieve adding new proxies.

There are some access violations in some tests due to this change. Pushing to code to another branch so that I can work on it later.