Automock to return proxy as type requested.
Updated the automock mock function to return the mock as the type requested. Previously this was giving an error during the query interface stage. Now the result should be the proxies virtual interface implementing the requested type. There is still an issue when the virtual interface is used. It is reporting that it does not implement the requested type. Another commit to come to fix this.