Update Delphi.Mocks.Proxy.pas
Fix for Mocks implementing multiple interfaces. When accessing the mock from different interfaces there will be an "Interface not supported" exception, while the real object implements al the interfaces and the mock seams not to. See code:

FBookAccount := TMock<IBookAccount>.Create;
FBookAccount.Implement<IPersistObject>;
FBookAccount.Implement<IModifiable>;
FBookAccount.Implement<IAsObjectInterface>;

// this worked:

a := FBookAccount;
b := a as IPersistObject;
c := a as IModifiable;
d := a as IAsObjectInterface;
i := a as IInterface;

// this did not work with the mock 
a := FBookAccount;
b := a as IPersistObject;
c := b as IModifiable;
d := c as IAsObjectInterface;
i := d as IInterface;